+++ src/Traits/Models/WrapQuery.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Php\Support\Laravel\Traits\Models;
 
-
 /**
  * Trait WrapQuery
  * @package Php\Support\Laravel\Traits\Models