+++ tests/TestClasses/Models/BaseModel.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace Php\Support\Laravel\Tests\TestClasses\Models;
+
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Foundation\Auth\User;
+use Php\Support\Laravel\Tests\TestClasses\Entity\Params;
+
+/**
+ * Class TestModel
+ * @package Php\Support\Laravel\Tests\Models
+ * @property boolean $enabled
+ * @property string $title
+ * @property Params $params
+ * @property array $config
+ * @property string $str
+ * @property string $str_empty
+ * @property int $int
+ * @property User $user
+ * @mixin Builder
+ */
+class BaseModel extends Model
+{
+    public $timestamps = false;
+
+    protected $table = 'base_table';
+
+}