+++ CHANGELOG.md
@@ -9,6 +9,13 @@ Check MD [online][check-online].
 
 ## [unreleased]
 
+## [0.9.0] - 2020-08-13
+
+### Added
+
+- Add trait `Modelable`
+- Add trait `ModelQueryable`
+
 ## [0.8.0] - 2020-07-08
 
 ### Added
@@ -109,7 +116,9 @@ Check MD [online][check-online].
 - The trait `CasterAttribute`: for custom class casts
 - The trait for DB/Migration `UUID`
 
-[unreleased]: https://github.com/efureev/laravel-support/compare/v0.7.0...HEAD
+[unreleased]: https://github.com/efureev/laravel-support/compare/v0.9.0...HEAD
+[0.8.0]: https://github.com/efureev/laravel-support/compare/v0.8.0...v0.9.0
+[0.8.0]: https://github.com/efureev/laravel-support/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/efureev/laravel-support/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/efureev/laravel-support/compare/v0.5.3...v0.6.0
 [0.5.3]: https://github.com/efureev/laravel-support/compare/v0.5.2...v0.5.3